--- datasets/bibtex/cityscapes.tex
@@ -1,6 +1,0 @@
-@inproceedings{Cordts2016Cityscapes,
-title={The Cityscapes Dataset for Semantic Urban Scene Understanding},
-author={Cordts, Marius and Omran, Mohamed and Ramos, Sebastian and Rehfeld, Timo and Enzweiler, Markus and Benenson, Rodrigo and Franke, Uwe and Roth, Stefan and Schiele, Bernt},
-booktitle={Proc. of the IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
-year={2016}
-}
