--- datasets/bibtex/facades.tex
@@ -1,7 +1,0 @@
-@INPROCEEDINGS{Tylecek13,
-  author = {Radim Tyle{\v c}ek, Radim {\v S}{\' a}ra},
-  title = {Spatial Pattern Templates for Recognition of Objects with Regular Structure},
-  booktitle = {Proc. GCPR},
-  year = {2013},
-  address = {Saarbrucken, Germany},
-}
