--- datasets/bibtex/handbags.tex
@@ -1,13 +1,0 @@
-@inproceedings{zhu2016generative,
-  title={Generative Visual Manipulation on the Natural Image Manifold},
-  author={Zhu, Jun-Yan and Kr{\"a}henb{\"u}hl, Philipp and Shechtman, Eli and Efros, Alexei A.},
-  booktitle={Proceedings of European Conference on Computer Vision (ECCV)},
-  year={2016}
-}
-
-@InProceedings{xie15hed,
-  author = {"Xie, Saining and Tu, Zhuowen"},
-  Title = {Holistically-Nested Edge Detection},
-  Booktitle = "Proceedings of IEEE International Conference on Computer Vision",
-  Year  = {2015},
-}
