--- datasets/bibtex/shoes.tex
@@ -1,14 +1,0 @@
-@InProceedings{fine-grained,
-  author = {A. Yu and K. Grauman},
-  title = {{F}ine-{G}rained {V}isual {C}omparisons with {L}ocal {L}earning},
-  booktitle = {Computer Vision and Pattern Recognition (CVPR)},
-  month = {June},
-  year = {2014}
-}
-
-@InProceedings{xie15hed,
-  author = {"Xie, Saining and Tu, Zhuowen"},
-  Title = {Holistically-Nested Edge Detection},
-  Booktitle = "Proceedings of IEEE International Conference on Computer Vision",
-  Year  = {2015},
-}
