--- models/models.py
@@ -1,29 +1,0 @@
-
-def create_model(opt):
-    model = None
-    print(opt.model)
-    if opt.model == 'cycle_gan':
-        assert(opt.dataset_mode == 'unaligned')
-        from .cycle_gan_model import CycleGANModel
-        model = CycleGANModel()
-    elif opt.model == 'pix2pix':
-        assert(opt.dataset_mode == 'aligned')
-        from .pix2pix_model import Pix2PixModel
-        model = Pix2PixModel()
-    elif opt.model == 'test':
-        assert(opt.dataset_mode == 'single')
-        from .test_model import TestModel
-        model = TestModel()
-    elif opt.model == 'gen_only':
-        assert(opt.dataset_mode == 'discrete')
-        from .generator_only_model import GeneratorOnlyModel
-        model = GeneratorOnlyModel()
-    elif opt.model == 'deeplab':
-        assert(opt.dataset_mode == 'discrete')
-        from .deeplab_model import DeeplabModel
-        model = DeeplabModel()
-    else:
-        raise ValueError("Model [%s] not recognized." % opt.model)
-    model.initialize(opt)
-    print("model [%s] was created" % (model.name()))
-    return model
