--- scripts/train_gen_only.sh
@@ -1,4 +1,0 @@
-#python train.py --dataroot ./datasets/cityscapes_AB --name city_gen_only --model gen_only --dataset_mode discrete --pool_size 50 --output_nc 20 --which_model_netG resnet_softmax_9blocks --batchSize 8 --continue_train
-#python train_gpu1.py --dataroot ./datasets/cityscapes_AB --name city_gen_only_resnet50 --model gen_only --dataset_mode discrete --output_nc 20 --which_model_netG resnet50_fcn --batchSize 1 --lr 0.0001 --lr_scheme lut --beta1 0.5 --niter 160 --continue_train --which_epoch 5
-#python train_gpu1.py --dataroot ./datasets/cityscapes_orig_AB --name city_gen_only_resnet9blocks --model gen_only --dataset_mode discrete --output_nc 20 --which_model_netG resnet50_fcn --batchSize 1 --lr 0.0002 --lr_scheme lut --beta1 0.5 --niter 160 --fineSize 896 --loadSize 1024 --ngf 128 --resize_or_crop crop_only --port 3430
-python train.py --dataroot ./datasets/cityscapes_572_AB --name city_gen_only_resnet9blocks --model gen_only --dataset_mode discrete --output_nc 20 --which_model_netG resnet_softmax_9blocks --batchSize 4 --lr 0.0002 --lr_scheme lut --beta1 0.5 --niter 160 --widthSize 512 --heightSize 256 --loadSize 572 --ngf 64 --resize_or_crop crop_only --port 9800
