+++ util/util.py
@@ -10,18 +10,19 @@ import collections
 # Converts a Tensor into a Numpy array
 # |imtype|: the desired type of the converted numpy array
 def tensor2im(image_tensor, imtype=np.uint8):
-    image_numpy = image_tensor[0].cpu().float().numpy()
+    image_numpy = image_tensor[0]
     #image_numpy = (np.transpose(image_numpy, (1, 2, 0)) + 1) / 2.0 * 255.0
     image_numpy = (np.transpose(image_numpy, (1, 2, 0))) * 255.0
     return image_numpy.astype(imtype)
 
 def tensor2lab(lab_tensor, label2color):
-    lab_numpy = lab_tensor[0].cpu().float().numpy()
-    if len(lab_numpy.shape) > 2:
-        seg_map = np.argmax(lab_numpy, axis=0) # pred
-    else:
-        seg_map = lab_numpy # gt
-    seg_map = seg_map.astype(np.int32).squeeze()
+    seg_map = lab_tensor[0]
+    #if len(lab_numpy.shape) > 2:
+    #    seg_map = np.argmax(lab_numpy, axis=0) # pred
+    #else:
+    #    seg_map = lab_numpy # gt
+    #seg_map = seg_map.astype(np.int32).squeeze()
+    seg_map = seg_map.astype(np.int32)
     assert(len(seg_map.shape) == 2)
     seg_image = label2color[seg_map].astype(np.uint8)
     return seg_image
@@ -37,7 +38,6 @@ def diagnose_network(net, name='network'):
         mean = mean / count
     print(name)
     print(mean)
-
 
 def save_image(image_numpy, image_path):
     image_pil = Image.fromarray(image_numpy)
